........
  r1493 | cstein | 2010-04-21 08:54:29 -0700 (Wed, 21 Apr 2010) | 5 lines
  
  Fixing type field mismatch in call.  This seems to fix a specific intermittent
  crash on 64bit platforms involving TIFFs.
........


git-svn-id: http://svn.openimageio.org/oiio/branches/RB-0.7@1495 31904243-a364-42ac-a3db-9305396518ba

+++ src/tiff.imageio/tiffinput.cpp
@@ -562,7 +562,7 @@ TIFFInput::readspec ()
     int iptcsize = 0;
     const void *iptcdata = NULL;
     if (TIFFGetField (m_tif, TIFFTAG_RICHTIFFIPTC, &iptcsize, &iptcdata)) {
-        std::vector<long> iptc ((long *)iptcdata, (long *)iptcdata+iptcsize);
+        std::vector<uint32> iptc ((uint32 *)iptcdata, (uint32 *)iptcdata+iptcsize);
         if (TIFFIsByteSwapped (m_tif))
             TIFFSwabArrayOfLong ((uint32*)&iptc[0], iptcsize);
         OpenImageIO::decode_iptc_iim (&iptc[0], iptcsize*4, m_spec);
